Make the crate callable as a library
This change moves most of the functionality in main.rs to lib.rs and
makes main() call it from there. The effect is that the functionality
of this crate is available as a library crate as well as built into
a binary.
